backend/requirements.txt
flask
flask-cors
requests
openai
python-dotenv
firebase-admin